Create rake task to move default banner image (#350)
* Create rake task to move default banner image

* Rubocop, my nemesis

+++ app/models/branding.rb
@@ -14,7 +14,7 @@ class Branding
     assets_path = Rails.root.join("public", "assets")
     orig_file_path = File.join(assets_path, "banner_image.jpg")
     # Move the original banner image to _old
-    FileUtils.mv(orig_file_path, File.join(assets_path, "banner_image_old.jpg"))
+    FileUtils.mv(orig_file_path, File.join(assets_path, "banner_image_old.jpg")) if File.exist?(orig_file_path)
     # Move the uploaded banner image to banner_image.jpg
     FileUtils.cp(temp_file_path, orig_file_path)
     # Get rid of the tempfile
